# openapi.yml
openapi: 3.0.3
info:
  title: ClubManager API
  description: >
    Specification for the ClubManager API,
    used by web and mobile clients to interact with the server.
  version: 0.1.0
...
components:
  securitySchemes:
  # I'm not sure if we are going to need BasicAuth depends how we handle Cognito
    BasicAuth:
      type: http
      scheme: basic
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - BearerAuth: [] 
paths:
...
  /api/bike/new:
    post:
      tags:
        - bike
      summary: Create a new bike
      description: "This endpoint requires an `Authorization: Bearer` header with a JWT ID token."
      requestBody:
        content:
          application/json:
            schema:
              # TODO
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                # TODO
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                # TODO
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                # TODO
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                # TODO
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                # TODO
  /api/bike/{bikeID}:
    parameters:
      - name: bikeID
        in: path
        required: true
        schema:
          type: integer
    get:
      tags:
        - bike
      summary: Get this bike
      description: "This endpoint requires an `Authorization: Bearer` header with a JWT ID token."
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                # TODO
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                # TODO
        404:
          description: Not Found
          content:
            application/json:
              schema:
                # TODO
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                # TODO
    patch:
      tags:
        - bike
      summary: Update this bike
      description: >
        The request body needs only contain the fields with data to be updated.
        If data in a field should not be updated, that field can be omitted from
        the request body.
        <br><br>
        This endpoint requires an `Authorization: Bearer` header with a JWT ID token.
      requestBody:
        content:
          application/json:
            schema:
              # TODO
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                # TODO
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                # TODO
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                # TODO
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                # TODO
        404:
          description: Not Found
          content:
            application/json:
              schema:
                # TODO
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                # TODO
    delete:
      tags:
        - bike
      summary: Delete this bike
      description: "This endpoint requires an `Authorization: Bearer` header with a JWT ID token."
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                # TODO
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                # TODO
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                # TODO
        404:
          description: Not Found
          content:
            application/json:
              schema:
                # TODO
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                # TODO
  /api/bike/list:
    get:
      tags:
        - bike
      summary: Get all bikes
      description: "This endpoint requires an `Authorization: Bearer` header with a JWT ID token."
      parameters:
        - name: membershipID
          in: query
          description: "Optional: pull bikes for a specific membership"
          schema:
            type: integer
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                # TODO
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                # TODO
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                # TODO

  # event
  /api/event/new:
    post:
      tags:
        - event
      summary: Create a new event
      description: "This endpoint requires an `Authorization: Bearer` header with a JWT ID token."
      requestBody:
        content:
          application/json:
            schema:
              # TODO
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                # TODO
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                # TODO
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                # TODO
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                # TODO
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                # TODO
  /api/event/{eventID}:
    parameters:
      - name: eventID
        in: path
        required: true
        schema:
          type: integer
    get:
      tags:
        - event
      summary: Get this event
      description: "This endpoint requires an `Authorization: Bearer` header with a JWT ID token."
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                # TODO
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                # TODO
        404:
          description: Not Found
          content:
            application/json:
              schema:
                # TODO
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                # TODO
    patch:
      tags:
        - event
      summary: Update this event
      description: >
        The request body needs only contain the fields with data to be updated.
        If data in a field should not be updated, that field can be omitted from
        the request body.
        <br><br>
        This endpoint requires an `Authorization: Bearer` header with a JWT ID token.
      requestBody:
        content:
          application/json:
            schema:
              # TODO
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                # TODO
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                # TODO
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                # TODO
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                # TODO
        404:
          description: Not Found
          content:
            application/json:
              schema:
                # TODO
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                # TODO
    delete:
      tags:
        - event
      summary: Delete this event
      description: "This endpoint requires an `Authorization: Bearer` header with a JWT ID token."
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                # TODO
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                # TODO
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                # TODO
        404:
          description: Not Found
          content:
            application/json:
              schema:
                # TODO
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                # TODO
  /api/event/list:
    get:
      tags:
        - event
      summary: Get all events
      description: "This endpoint requires an `Authorization: Bearer` header with a JWT ID token."
      parameters:
        - name: Range
          in: header
          description: |
            Optional: filter events by a specific timeframe.
            Format is one of:
            - `YYYYMMDD-YYYYMMDD`
            - `YYYYMMDD-`
            - `-YYYYMMDD`
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                # TODO
        206:
          description: Partial Content
          content:
            application/json:
              schema:
                # TODO
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                # TODO
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                # TODO
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                # TODO
...
  /api/billing/yearlyWorkPointThreshold:
    get:
      tags:
        - billing
      summary: Get a year's work point threshold for completely subsidizing dues
      description: "This endpoint requires an `Authorization: Bearer` header with a JWT ID token."
      parameters:
        - name: year
          in: query
          description: "Optional: pull the threshold for a specific year"
          schema:
            type: integer
            default: the current year
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                # TODO
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                # TODO
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                # TODO
  /api/billing/list:
    get:
      tags:
        - billing
      summary: Get all bills by payment status or year
      description: "This endpoint requires an `Authorization: Bearer` header with a JWT ID token."
      parameters:
        - name: paymentStatus
          in: query
          description: "Optional: filter bills by a specific payment status"
          schema:
            type: string
            enum:
              - paid
              - outstanding
        - name: year
          in: query
          description: "Optional: filter bills by a specific year"
          schema:
            type: integer
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                # TODO
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                # TODO
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                # TODO
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                # TODO
  /api/billing:
    post:
      tags:
        - billing
      summary: Calculate and distribute all bills for the year
      description: "This endpoint requires an `Authorization: Bearer` header with a JWT ID token."
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                # TODO
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                # TODO
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                # TODO
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                # TODO
  /api/billing/{membershipID}:
    parameters:
      - name: membershipID
        in: path
        required: true
        schema:
          type: integer
    get:
      tags:
        - billing
      summary: Get bills for this membership
      description: "This endpoint requires an `Authorization: Bearer` header with a JWT ID token."
      parameters:
        - name: year
          in: query
          description: "Optional: filter bills by a specific year"
          schema:
            type: integer
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                # TODO
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                # TODO
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                # TODO
        404:
          description: Not Found
          content:
            application/json:
              schema:
                # TODO
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                # TODO
    post:
      tags:
        - billing
      summary: Pay this membership's outstanding balance
      description: "This endpoint requires an `Authorization: Bearer` header with a JWT ID token."
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                # TODO
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                # TODO
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                # TODO
        404:
          description: Not Found
          content:
            application/json:
              schema:
                # TODO
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                # TODO